requirements.txt
langchain
wikipedia
langchain-community
langchain-core
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
flask